requirements.txt
async-generator==1.10
attrs==22.1.0
awscli==1.26.4
botocore==1.28.4
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
contourpy==1.0.6
cycler==0.11.0
dnspython==2.2.1
docutils==0.16
dominate==2.7.0
email-validator==1.3.0
exceptiongroup==1.0.0rc9
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.38.0
git-remote-codecommit==1.16
greenlet==1.1.3
gunicorn==20.1.0
h11==0.14.0
idna==3.4
imageio==2.22.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
mnist==0.2.2
mysql-connector==2.2.9
networkx==2.8.8
numpy==1.23.4
outcome==1.2.0
packaging==21.3
Pillow==9.3.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
PyMySQL==1.0.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==5.4.1
requests==2.28.1
rsa==4.7.2
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.41
threadpoolctl==3.1.0
tifffile==2022.10.10
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
visitor==0.1.3
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
